completes workspace switching, fixes some edge case auth bugs
