+++ backend/app/controllers/api/v1/team_controller.rb
@@ -4,6 +4,10 @@ module Api
       def users
         render json: current_workspace.users, status: :ok
       end
+
+      def workspace_members
+        render json: current_workspace.workspace_members, each_serializer: WorkspaceMemberSerializer, status: :ok
+      end
     end
   end
 end
