+++ backend/app/controllers/api/v1/workspaces_controller.rb
@@ -4,10 +4,9 @@ module Api
       def update
         workspace = current_workspace
         if workspace.update(workspace_params)
-          user = current_user
           # reset the token with the updated workspace
           # this also resets the expires_at, probably should have a better way to do this but ok for now
-          log_user_out
+          user = log_user_out
           token = log_user_in(user, workspace)
           render json: { workspace: workspace, auth_token: token }, status: :ok
         else
@@ -17,7 +16,8 @@ module Api
 
       def update_current_workspace
         new_workspace = current_user.workspaces.find(params[:workspace_id])
-        token = log_user_in(current_user, new_workspace)
+        user = log_user_out
+        token = log_user_in(user, new_workspace)
         render json: { workspace: new_workspace, auth_token: token }, status: :ok
       end
