+++ dashboard/src/app/(authenticated)/team/page.js
@@ -1,27 +1,25 @@
 'use client';
 
-// import { formattedDate } from '@/lib/utils';
 import CopyToClipboard from 'react-copy-to-clipboard';
-import LoadingSpinner from '@/components/LoadingSpinner';
+import InviteModal from '@/components/TeamManagement/NewInvitationModal';
+import RemoveWorkspaceMemberModal from '@/components/TeamManagement/RemoveWorkspaceMemberModal';
 import { Menu, Transition } from '@headlessui/react';
-import Modal from '@/components/utils/Modal';
 import { SwishjamAPI } from "@/lib/api-client/swishjam-api";
 import { useAuthData } from '@/hooks/useAuthData'
-import { UserPlusIcon, EnvelopeIcon, ClipboardDocumentIcon, CheckCircleIcon, EllipsisVerticalIcon, TrashIcon } from '@heroicons/react/24/outline';
+import { UserPlusIcon, ClipboardDocumentIcon, EllipsisVerticalIcon, TrashIcon } from '@heroicons/react/24/outline';
 import { useState, useEffect, Fragment } from 'react';
 
-const formattedDate = date => date;
 const classNames = (...classes) => classes.filter(Boolean).join(' ')
 
 export default function Team() {
   const { userId: currentUserId } = useAuthData()
   const [inviteModalIsOpen, setInviteModalIsOpen] = useState(false);
-  const [users, setUsers] = useState();
-  const [userToDisplayInRemoveModal, setUserToDisplayInRemoveModal] = useState();
+  const [workspaceMembers, setWorkspaceMembers] = useState();
+  const [workspaceMemberToDisplayInRemoveModal, setWorkspaceMemberToDisplayInRemoveModal] = useState();
 
   useEffect(() => {
-    setUsers();
-    SwishjamAPI.Team.users().then(setUsers)
+    setWorkspaceMembers();
+    SwishjamAPI.Team.workspaceMembers().then(setWorkspaceMembers)
   }, [])
 
   return (
@@ -29,15 +27,17 @@ export default function Team() {
       <InviteModal
         isOpen={inviteModalIsOpen}
         onClose={() => setInviteModalIsOpen(false)}
-        onInviteSent={userInvite => setUsers([...users, { ...userInvite, email: userInvite.invited_email, status: 'pending' }])}
+      // we're not currently displaying pending invites here
+      // onInviteSent={userInvite => setWorkspaceMembers([...workspaceMembers, { ...userInvite, email: userInvite.invited_email, status: 'pending' }])}
       />
-      <RemoveUserModal
-        isOpen={!!userToDisplayInRemoveModal}
-        userId={userToDisplayInRemoveModal?.id}
-        userEmail={userToDisplayInRemoveModal?.email}
-        onClose={() => setUserToDisplayInRemoveModal()}
-        onRemoveUser={user => setUsers(users.filter(u => u.id !== user.id))}
-      />
+      {workspaceMemberToDisplayInRemoveModal && (
+        <RemoveWorkspaceMemberModal
+          workspaceMemberId={workspaceMemberToDisplayInRemoveModal.id}
+          userEmail={workspaceMemberToDisplayInRemoveModal.user.email}
+          onClose={() => setWorkspaceMemberToDisplayInRemoveModal()}
+          onRemoveWorkspaceMember={workspaceMemberId => setWorkspaceMembers(workspaceMembers.filter(wm => wm.id !== workspaceMemberId))}
+        />
+      )}
       <main className="mx-auto max-w-7xl px-4 sm:px-6 lg:px-8 mb-8 mt-12">
         <div className='my-8 grid grid-cols-2'>
           <h1 className="text-lg font-medium text-gray-700 mb-0">Team Management</h1>
@@ -69,15 +69,15 @@ export default function Team() {
             </tr>
           </thead>
           <tbody className="divide-y divide-gray-200 bg-white">
-            {users
+            {workspaceMembers
               ? (
-                users.map(({ id, email, created_at, invite_token, status = 'accepted' }) => (
+                workspaceMembers.map(({ id, user, created_at, invite_token, status = 'accepted' }) => (
                   <tr key={id} className='hover:bg-gray-50'>
                     <td className="cursor-default whitespace-nowrap py-3.5 px-3 text-sm text-gray-700 max-w-xs truncate overflow-hidden">
-                      {email}
+                      {user.email}
                     </td>
                     <td className="cursor-default whitespace-nowrap py-3.5 px-3 text-sm text-gray-700 max-w-xs truncate overflow-hidden">
-                      {status === 'accepted' && formattedDate(created_at)}
+                      {status === 'accepted' && new Date(created_at).toLocaleDateString('en-US', { month: 'long', day: 'numeric', year: 'numeric' })}
                     </td>
                     <td className="cursor-default whitespace-nowrap py-3.5 px-3 text-sm text-gray-700 max-w-xs truncate overflow-hidden">
                       <div className={`inline-flex items-center rounded-md  px-2 py-1 text-xs font-medium ring-1 ring-inset ${status === 'accepted' ? 'bg-green-50 text-green-700 ring-green-600/20' : 'bg-orange-50 text-orange-700 ring-orange-600/20'}`}>
@@ -86,8 +86,8 @@ export default function Team() {
                     </td>
                     <td className="cursor-default whitespace-nowrap py-3.5 px-3 text-gray-700 text-sm max-w-xs">
                       <div className="py-1 flex items-center justify-end">
-                        {id !== currentUserId && (
-                          <UserDropdown status={status} inviteToken={invite_token} onRemoveUserClick={() => setUserToDisplayInRemoveModal({ id, email })} />
+                        {user.id !== currentUserId && (
+                          <UserDropdown status={status} inviteToken={invite_token} onRemoveUserClick={() => setWorkspaceMemberToDisplayInRemoveModal({ id, user })} />
                         )}
                       </div>
                     </td>
@@ -188,174 +188,3 @@ const UserDropdown = ({ status, inviteToken, onRemoveUserClick }) => {
   )
 }
 
-const RemoveUserModal = ({ userId, userEmail, isOpen, onClose, onRemoveUser }) => {
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const [error, setError] = useState();
-  const [showSuccessMessage, setShowSuccessMessage] = useState(false);
-
-  const removeUser = async e => {
-    e.preventDefault();
-    setIsSubmitting(true);
-    const { error } = await SwishjamAPI.post('/api/organization-users/delete', { userId });
-    if (error) {
-      setIsSubmitting(false);
-      setError(error);
-    } else {
-      await onRemoveUser({ id: userId, email: userEmail });
-      setShowSuccessMessage(true);
-    }
-  }
-
-  return (
-    <Modal
-      isOpen={isOpen}
-      onClose={() => {
-        onClose();
-        setTimeout(() => {
-          setIsSubmitting(false);
-          setError();
-          setShowSuccessMessage(false);
-        }, 500);
-      }}
-    >
-      {showSuccessMessage
-        ? (
-          <div className='px-4 py-8 text-center'>
-            <h3 className='text-lg'>{userEmail} no longer has access to your Swishjam team.</h3>
-          </div>
-        ) : (
-          <div className='p-4 text-center'>
-            <h3 className='text-lg mb-2'>Are you sure you'd like to remove {userEmail} from your team?</h3>
-            {error && <p className='text-sm text-red-600 mb-2'>{error}</p>}
-            <div className='grid grid-cols-2 gap-x-4'>
-              <button
-                className='mt-4 flex items-center justify-center py-2 px-4 border border-gray-200 rounded-md shadow-sm text-sm font-medium text-gray-700 focus:outline-none focus:ring-2 bg-white hover:bg-gray-100 focus:ring-offset-2 focus:ring-gray-100'
-                onClick={() => {
-                  onClose();
-                  setTimeout(() => {
-                    setIsSubmitting(false);
-                    setError();
-                    setShowSuccessMessage(false);
-                  }, 500);
-                }}
-              >
-                Nevermind
-              </button>
-              <button
-                onClick={removeUser}
-                className={`mt-4 flex items-center justify-center py-2 px-4 border border-red-600 rounded-md shadow-sm text-sm font-medium text-white focus:outline-none focus:ring-2 bg-red-500 hover:bg-red-600 focus:ring-offset-2 focus:ring-gray-600 ${isSubmitting ? 'opacity-50 cursor-not-allowed' : ''}`}
-                disabled={isSubmitting}
-              >
-                {isSubmitting ? 'Removing...' : 'Remove'}
-              </button>
-            </div>
-          </div>
-        )
-      }
-    </Modal>
-  )
-}
-
-const InviteModal = ({ isOpen, onClose, onInviteSent }) => {
-  const [email, setEmail] = useState('');
-  const [isSubmitting, setIsSubmitting] = useState(false);
-  const [error, setError] = useState();
-  const [successMessage, setSuccessMessage] = useState();
-  const [userInviteToken, setUserInviteToken] = useState();
-  const [copySuccessMessage, setCopySuccessMessage] = useState();
-
-  const inviteUser = async e => {
-    e.preventDefault();
-    setUserInviteToken();
-    setIsSubmitting(true);
-    setError();
-    setSuccessMessage();
-    const response = await SwishjamAPI.WorkspaceInvitations.create({ email });
-    debugger;
-    setIsSubmitting(false);
-    if (response.error) {
-      setError(response.error);
-      setSuccessMessage();
-    } else {
-      setUserInviteToken(response.invite_token);
-      setSuccessMessage(`Invitation sent to ${response.invited_email}.`);
-      setEmail('');
-      onInviteSent && onInviteSent(response);
-    }
-  }
-
-  return (
-    <Modal
-      isOpen={isOpen}
-      onClose={() => {
-        onClose();
-        setTimeout(() => {
-          setEmail('');
-          setError();
-          setUserInviteToken();
-          setSuccessMessage();
-        }, 500)
-      }}
-    >
-      {userInviteToken
-        ? (
-          <>
-            <div className='text-center'>
-              <CheckCircleIcon className='h-12 w-12 text-green-500 mx-auto mb-2' />
-              <h2 className='text-md font-medium text-gray-900 mb-4'>{successMessage}</h2>
-              <div className='grid grid-cols-2 justify-center gap-x-4'>
-                <CopyToClipboard
-                  text={`${window.location.origin}/invitation/${userInviteToken}`}
-                  onCopy={() => {
-                    setCopySuccessMessage('Copied!')
-                    setTimeout(() => setCopySuccessMessage(), 5_000);
-                  }}
-                >
-                  <button className={`mt-4 flex items-center justify-center py-2 px-4 border border-gray-200 rounded-md shadow-sm text-sm font-medium text-gray-700 focus:outline-none focus:ring-2 bg-white hover:bg-gray-100 focus:ring-offset-2 focus:ring-gray-100`}>
-                    <ClipboardDocumentIcon className='h-4 w-4 inline-block mr-1' />
-                    {copySuccessMessage || 'Copy invitation link'}
-                  </button>
-                </CopyToClipboard>
-                <button
-                  type="button"
-                  className={`mt-4 flex items-center justify-center py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white focus:outline-none focus:ring-2 bg-swishjam hover:bg-swishjam-dark focus:ring-offset-2 focus:ring-swishjam-dark`}
-                  onClick={() => {
-                    setUserInviteToken();
-                    setSuccessMessage();
-                  }}
-                >
-                  Invite another teammate
-                </button>
-              </div>
-            </div>
-          </>
-        ) : (
-          <>
-            <h2 className='text-md font-medium text-gray-900 mb-4'>Invite a teammate</h2>
-            <form onSubmit={inviteUser}>
-              <div className='mt-2'>
-                <label className='block text-sm font-medium text-gray-700'>Teammate's Email</label>
-                <input className='input mt-1' placeholder='johnny@gmail.com' type='email' value={email} onChange={e => setEmail(e.target.value)} />
-              </div>
-              {error && <div className='text-sm text-center text-red-500 mt-2'>{error}</div>}
-              <button
-                type="submit"
-                className={`w-full mt-4 flex items-center justify-center py-2 px-4 border border-transparent rounded-md shadow-sm text-sm font-medium text-white focus:outline-none focus:ring-2 ${isSubmitting ? 'bg-gray-400' : 'bg-swishjam hover:bg-swishjam-dark'}`}
-                disabled={isSubmitting}
-              >
-                {isSubmitting ?
-                  <div className="h-6"><LoadingSpinner size={6} color='white' /></div>
-                  : (
-                    <>
-                      <EnvelopeIcon className='h-4 w-4 inline-block mr-1' />
-                      Send Invite
-                    </>
-                  )
-                }
-              </button>
-            </form>
-          </>
-        )}
-    </Modal>
-  )
-}
