+++ dashboard/src/lib/api-client/auth.js
@@ -8,7 +8,7 @@ export class Auth extends Base {
   static async login({ email, password }) {
     return this._post('/auth/login', { email, password });
   }
-  
+
   static async register({ email, password, companyUrl, inviteCode }) {
     return await this._post('/auth/register', { email, password, company_url: companyUrl, invite_code: inviteCode })
   }
